% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendarHeat.R
\name{calendarHeat}
\alias{calendarHeat}
\title{日历热图}
\usage{
calendarHeat(dates, values, ncolors = 99, color = "r2g",
  varname = "Values", week_label = c("周日", "周一", "周二", "周三",
  "周四", "周五", "周六"), month_label = paste0(1:12, "月"),
  date.form = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{dates}{日期}

\item{values}{数据}

\item{ncolors}{颜色}

\item{color}{颜色}

\item{varname}{变量}

\item{week_label}{周标签}

\item{month_label}{月标签}

\item{date.form}{日期来源}

\item{...}{其他}
}
\value{
返回值
}
\description{
日历热图
}
\examples{
calendarHeat(bb);
}
